% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/different.tan-package.R
\docType{package}
\name{different.tan}
\alias{different.tan}
\title{different.tan}
\description{
different.tan: R code for the analysis of 
vehicle emissions remote sensing system (VERSS) 
data
}
\section{different.tan}{

\code{different.tan} was initially developed as part 
of work on a series of projects collecting and 
analyzing VERSS data.

The name 'different.tan' refers to Don Stedman, a pioneer 
in field, who died in 2016: 
\href{https://magazine-archive.du.edu/campus-community/chemistry-professor-donald-stedman-dies-lung-cancer/}{[DON]}
}

\references{
We gratefully acknowledge the huge contributions of the R Core Team 
and numerous contributors in developing and maintaining R:

R Development Core Team (2011). R: A language and environment for 
statistical computing. R Foundation for Statistical Computing, 
Vienna, Austria. ISBN 3-900051-07-0, URL http://www.R-project.org/.

EDAR study references to add...
}
\author{
Karl Ropkins, Francis Pope, David Beddows, Omid Ghaffarpasand; 
(maintainer) Karl Ropkins <k.ropkins@its.leeds.ac.uk>
}
