% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_edardata.R
\name{import_edardata}
\alias{import_edardata}
\title{import_edardata}
\usage{
import_edardata(edardata = NULL, simplify.names = TRUE, ...)
}
\arguments{
\item{edardata}{Character, data of set name (see below)}

\item{simplify.names}{Logical, if \code{TRUE}, simplifies the column names 
(converting to lower case, changes underscores to dots, etc), and replaces 
converts from HEATs m[n]_name format so analytes are identified in column 
names.}

\item{...}{Currently ignored.}
}
\value{
\code{import_edardata} attempts to match \code{edardata} and, if found, 
returns it as a \code{data.frame}.  

The function uses base \code{R} code, \code{\link{download.file}} to 
temporarily download data sets and \code{\link{readxl::read_xlsx}}  
to read data in. The \code{simplify.names=TRUE} argument is used 
to make data column names a little r-friendlier.
}
\description{
Function imports EDAR data sets from the West Lothian (See below).
}
\details{
The function imports data from the 
\href{https://www.westlothian.gov.uk/article/45802/Real-Time-Vehicle-Emissions-pilot-project-March-2017}{Edinburgh and West Lothian Real Time Vehicle Emissions Stduy 2017}. 

\tabular{lll}{
 dataset \tab datatype \tab ref\cr

edin2017     \tab Edinburgh EDAR 2017 \tab [ref omid's paper?]                                         \cr
brox2017     \tab Edinburgh EDAR 2017 \tab [ref omid's paper?]                                         \cr
}
}
\note{
In-development: code may be subject to significant change.
}
\examples{
#to do...

}
\references{
For \code{read_xlsx} package and \code{read_xlsx}: 

 Wickham H, Bryan J (2022). _readxl: Read Excel Files_. 
 R package version 1.4.0, <https://CRAN.R-project.org/package=readxl>.

For general information about EDAR:

[my ref]?
}
\seealso{
\code{\link{readxl::read_xlsx}}.
}
