% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_featdata.R
\name{import_featdata}
\alias{import_featdata}
\title{import_featdata}
\usage{
import_featdata(featdata = NULL, simplify.names = TRUE, ...)
}
\arguments{
\item{featdata}{Character, data set name (see below)}

\item{simplify.names}{Logical, if \code{TRUE}, simplifies the column names of the 
found data set by converting them to lower case and changing underscores to dots.}

\item{...}{Currently ignored.}
}
\value{
\code{import_featdata} attempts to match \code{featdata} and, if found, 
returns it as a \code{data.frame}.  

The function uses base \code{R} code and \code{\link{foreign::read.dbf}} to download, 
read and tidy data. The \code{as.is=TRUE} argument is applied when reading data so, 
for example, \code{date} and \code{time} columns are imported as \code{character} 
vectors rather than \code{factors}.
}
\description{
Function imports data from the Denver FEAT data archive (See below).
}
\details{
The function imports data from the 
\href{https://digitalcommons.du.edu/feat/}{FEAT data archive}. 

\tabular{lll}{
 dataset \tab datatype \tab ref\cr

Tucson 94     \tab light duty vehicles \tab 10.1021/Es950191j                                         \cr
 Phnx 98       \tab light duty vehicles \tab 10.1021/es702413b                                         \cr
 Phnx 99       \tab light duty vehicles \tab 10.1021/es702413b                                         \cr
 AZ 2000       \tab light duty vehicles \tab 10.1021/es702413b                                         \cr
 AZ 2002       \tab light duty vehicles \tab 10.1021/es702413b                                         \cr
 AZ 2004       \tab light duty vehicles \tab 10.1021/es702413b                                         \cr
 AZ 2006       \tab light duty vehicles \tab 10.1021/es702413b                                         \cr
 Fresno 08     \tab light duty vehicles \tab 10.1021/acs.est.0c05433                                   \cr
 Fresno 21     \tab light duty vehicles \tab 10.1021/acs.est.0c05433                                   \cr
 Cal 89        \tab light duty vehicles \tab 10.1126/science.268.5213.991                              \cr
 Cal 91        \tab light duty vehicles \tab 10.1126/science.268.5213.991                              \cr
 LA710 99      \tab light duty vehicles \tab -                                                         \cr
 Lynwd105 2018 \tab light duty vehicles \tab 10.1080/10962247.2019.1611677                             \cr
 Lynwd710 2018 \tab light duty vehicles \tab 10.1080/10962247.2019.1611677                             \cr
 LA 99         \tab light duty vehicles \tab 10.1021/es026340x                                         \cr
 LA 00         \tab light duty vehicles \tab 10.1021/es026340x                                         \cr
 LA 01         \tab light duty vehicles \tab 10.1021/es026340x                                         \cr
 Smnto 99      \tab light duty vehicles \tab -                                                         \cr
 Sanjose 99    \tab light duty vehicles \tab 10.1021/acs.est.0c05433; 10.1021/acs.est.0c05433          \cr
 Sanjose 08    \tab light duty vehicles \tab 10.1021/acs.est.0c05433; 10.1021/acs.est.0c05433          \cr
 Vannuys 10    \tab light duty vehicles \tab 10.1080/10962247.2012.699015; 10.1080/10962247.2012.699016\cr
 Labrea 01     \tab light duty vehicles \tab 10.1080/10962247.2020.1869121                             \cr
 Labrea 03     \tab light duty vehicles \tab 10.1021/es702413b                                         \cr
 Labrea 05     \tab light duty vehicles \tab 10.1021/acs.est.0c05433                                   \cr
 Labrea 08     \tab light duty vehicles \tab 10.1021/es5043518                                         \cr
 Labrea 13     \tab light duty vehicles \tab 10.1021/acs.est.5b02392                                   \cr
 Labrea 15     \tab light duty vehicles \tab 10.1021/acs.est.6b00717                                   \cr
 Labrea 18     \tab light duty vehicles \tab 10.1021/acs.est.0c05433                                   \cr
 Labrea 99     \tab light duty vehicles \tab 10.1080/10962247.2019.1611677                             \cr
 Peralta 97    \tab heavy duty vehicles \tab 10.1021/Es001533a                                         \cr
 Peralta 08    \tab heavy duty vehicles \tab 10.1021/es202392g                                         \cr
 Peralta 09    \tab heavy duty vehicles \tab 10.1021/es202392g                                         \cr
 Peralta 10    \tab heavy duty vehicles \tab 10.1021/es202392g                                         \cr
 Peralta 12    \tab heavy duty vehicles \tab 10.1021/es401487b                                         \cr
 Peralta 17    \tab heavy duty vehicles \tab 10.1021/acs.est.8b03994                                   \cr
 laport 08     \tab heavy duty vehicles \tab 10.1021/es202392g, 10.1021/es401487b                      \cr
 laport 09     \tab heavy duty vehicles \tab 10.1021/es202392g, 10.1021/es401487b                      \cr
 laport 10     \tab heavy duty vehicles \tab 10.1021/es202392g, 10.1021/es401487b                      \cr
 laport 12     \tab heavy duty vehicles \tab 10.1021/es202392g, 10.1021/es401487b                      \cr
 coors 99      \tab heavy duty vehicles \tab 10.1021/Es001533a                                         \cr
 coors 05      \tab heavy duty vehicles \tab 10.1021/es060989a                                         \cr
 dumont 99     \tab heavy duty vehicles \tab 10.1021/Es001533a                                         \cr
 dumont 05     \tab heavy duty vehicles \tab 10.1021/es060989a                                         \cr
 uri 98        \tab heavy duty vehicles \tab 10.1021/Es001533a                                         \cr
 pohouston 09  \tab heavy duty vehicles \tab -                                                         \cr
 sanmarcos 98  \tab heavy duty vehicles \tab 10.1021/Es001533a                                         \cr
 utah 20       \tab heavy duty vehicles \tab 10.1021/Es001533a                                         \cr
 cottonwood 13 \tab heavy duty vehicles \tab 10.1021/es505534e                                         \cr
 cottonwood 15 \tab heavy duty vehicles \tab 10.1021/acs.est.6b06172                                   \cr
 cottonwood 17 \tab heavy duty vehicles \tab 10.1021/acs.est.8b00621                                   \cr
 pola 13       \tab heavy duty vehicles \tab 10.1021/es505534e                                         \cr
 pola 15       \tab heavy duty vehicles \tab 10.1021/acs.est.6b06172                                   \cr
 pola 17       \tab heavy duty vehicles \tab 10.1021/acs.est.8b00621                                   
}
}
\note{
In-development: code may be subject to significant change; Currently, only some 
data sets can be downloaded from the FEAT archive. (See Details above for list of 
data sets available using \code{import_featdata}; please check main archive, if you 
are after any other FEAT data not listed here.)
}
\examples{
#to do...

}
\references{
For \code{foreign} package and \code{read.dbf}: 

R Core Team (2022). _foreign: Read Data Stored by 'Minitab', 'S', 'SAS', 'SPSS',
'Stata', 'Systat', 'Weka', 'dBase', ..._. R package version 0.8-82, 
<https://CRAN.R-project.org/package=foreign>.

For general information about FEAT:

D.H. Stedman and G.A. Bishop. Measuring the Emissions of Passing Cars,  
Acc. Chem. Res., 29:489-495, 1996. 

D.A. Burgard, G.A. Bishop, R.S. Stadtmuller, T.R. Dalton and D.H. Stedman, 
Spectroscopy Applied to On-Road Mobile Source Emissions.
Appl. Spectrosc., 60:5:135A-148A, 2006.

And, for specific campaigns/data sets, please reference using study 
reference(s), listed as associated DOI(s) in Details above.
}
\seealso{
\code{\link{foreign::read.dbf}}.
}
