% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Pareto.R
\name{plot_Pareto}
\alias{plot_Pareto}
\title{plot_Pareto}
\usage{
plot_Pareto(
  y,
  groups = NULL,
  cond = NULL,
  data = NULL,
  x = NULL,
  ...,
  panel = panel.ParetoPlot,
  drop.nas = TRUE,
  Pareto = "normal.cum.Pareto"
)
}
\arguments{
\item{y}{(numeric vector), measurements to plotted as PartoPlot Y-axis terms. In VERSS 
analysis these are typically emission measurements from passing vehicles, but can be 
any data-series where the frequency of occurrence of \code{y} is of interest.}

\item{groups}{(optional) sub-sampling term for grouping measurements within a plot.}

\item{cond}{(optional) sub-sampling term for plotting data as separate panels.}

\item{data}{(optional, data.frame or similar), source of \code{y}, \code{groups},
\code{cond}, etc.}

\item{x}{(optional, number vector), by default, the index of \code{y} values, lowest 
to highest.}

\item{...}{Passed to lattice and loa plot functions (and used if possible).}

\item{panel}{lattice plot argument}

\item{drop.nas}{Remove NAs before plotting.}

\item{Pareto}{(Character) Pareto plot type, options: 'Pareto' standard Pareto; 
'cum.Pareto' cumulative Pareto; 'normal.cum.Pareto' cumulative Pareto expressed 
as a proportion; and, 'percent.cum.Pareto' cumulative Pareto expressed 
as a percentage.}
}
\value{
\code{lattice} object plot.
}
\description{
In-development plots functions intended for use with VERSS data.
}
