% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ParetoEnrichment.R
\name{plot_ParetoEnrichment}
\alias{plot_ParetoEnrichment}
\title{Pareto Enrichment Plots}
\usage{
plot_ParetoEnrichment(
  x,
  y = NULL,
  cut = 0.8,
  data = NULL,
  ...,
  min.bin = 1,
  drop.nas = TRUE,
  order = FALSE,
  level.nmax = NULL
)
}
\arguments{
\item{x}{[to be added]}

\item{y}{[to be added]}

\item{cut}{[to be added]}

\item{data}{[to be added]}

\item{...}{[to be added]}

\item{min.bin}{[to be added]}

\item{drop.nas}{[to be added]}

\item{order}{[to be added]}

\item{level.nmax}{[to be added]}
}
\description{
[to be added]
}
